Ultralytics migration from yolo 8 to 11
